Implement vehicle routes

+++ model/vehicle.go
@@ -1,9 +1,9 @@
 package model
 
 type Vehicle struct {
-	Id           int64  `json:"id" validate:"required"`
+	Id           int    `json:"id" validate:"required"`
 	Manufacturer string `json:"manufacturer" validate:"required"`
 	Model        string `json:"model" validate:"required"`
 	Vin          string `json:"vin" validate:"required,min=17,max=17"`
-	Performance  int32  `json:"performance" validate:"max=999"`
+	Performance  int    `json:"performance" validate:"max=999"`
 }
